dev5/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
Werkzeug==3.0.1
yt-dlp==2024.03.10
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
requests==2.31.0
cryptography==42.0.2
eventlet==0.35.2